trimming up the sdk by modularizing compnonents
broke out the sd card functionality from sensors. everything
still compiles, but the _output_buffer is still taking up
far too much space in ArdusatSensors. no real improvement yet

for acceleration example file:
ArdusatSKD: 16060 bytes of program, 1424 bytes of dynamic memory
ArdusatLogging: 16002 bytes of program, 1878 bytes of dynamic memory
ArdusatSensors: 16002 bytes of program, 1878 bytes of dynamic memory
