+++ ArdusatSensors.h
@@ -1,25 +1,20 @@
 /**
- * @file   ArdusatSDK.h
+ * @file   ArdusatSensors.h
  * @Author Ben Peters (ben@ardusat.com)
  * @date   December 3, 2014
- * @brief  ArdusatSDK generic sensor reading & configuration for Space Kit Sensors
+ * @brief  ArdusatSensors generic sensor reading & configuration for Space Kit Sensors
  */
-#ifndef ARDUSATSDK_H_
-#define ARDUSATSDK_H_
+#ifndef ARDUSATSENSORS_H_
+#define ARDUSATSENSORS_H_
 
 #include <Arduino.h>
-#include <utility/SdFat.h>
 #include <utility/drivers.h>
 #include <avr/pgmspace.h>
-#include <utility/SdVolume.h>
-#include <utility/MemoryFree.h>
-#include <utility/BinaryDataFmt.h>
-#include <utility/serial.h>
-#include <utility/RTClib.h>
 
 #define prog_char const char PROGMEM
 
 extern bool ARDUSAT_SHIELD;
+
 /**
  * Unique numeric id for each physical sensor
  */
@@ -58,19 +53,19 @@ typedef enum {
  * Data types for cells in data structures.
  */
 // schema for these constants is 			0000XXLL XX for type, LL for len-1
-#define DATA_CELLTYPE_HEX	0x00
+#define DATA_CELLTYPE_HEX	    0x00
 #define DATA_CELLTYPE_HEX8		0x00	//	00000000
 #define DATA_CELLTYPE_HEX16		0x01	//	00000001
 #define DATA_CELLTYPE_HEX24		0x02	//	00000010
 #define DATA_CELLTYPE_HEX32		0x03	//	00000011
 
-#define DATA_CELLTYPE_INT	0x04
+#define DATA_CELLTYPE_INT	    0x04
 #define DATA_CELLTYPE_INT8		0x04	//	00000100
 #define DATA_CELLTYPE_INT16		0x05	//	00000101
 #define DATA_CELLTYPE_INT24		0x06	//	00000110
 #define DATA_CELLTYPE_INT32		0x07	//	00000111
 
-#define DATA_CELLTYPE_UINT	0x08
+#define DATA_CELLTYPE_UINT	  0x08
 #define DATA_CELLTYPE_UINT8		0x08	//	00001000
 #define DATA_CELLTYPE_UINT16	0x09	//	00001001
 #define DATA_CELLTYPE_UINT24	0x0A	//	00001010
@@ -80,12 +75,12 @@ typedef enum {
 #define DATA_CELLTYPE_FLOAT		0x0F	//  00001111
 
 // utility constants
-#define DATA_CELLTYPE_LONGINT	0x07	// (INT4)
+#define DATA_CELLTYPE_LONGINT	  0x07	// (INT4)
 #define DATA_CELLTYPE_SHORTINT	0x05	// (INT2)
-#define DATA_CELLTYPE_INTEGER	0x05	// (SHORTINT) (INT2)
+#define DATA_CELLTYPE_INTEGER	  0x05	// (SHORTINT) (INT2)
 #define DATA_CELLTYPE_ULONGINT	0x0B	// (UINT4)
 #define DATA_CELLTYPE_USHORTINT	0x09	// (UINT2)
-#define DATA_CELLTYPE_BYTE		0x00	// (HEX1)
+#define DATA_CELLTYPE_BYTE		  0x00	// (HEX1)
 
 
 /**
@@ -219,53 +214,8 @@ const char * uvlightToJSON(const char *sensorName, uvlight_t * input);
 const char * orientationToJSON(const char *sensorName, orientation_t * input);
 const char * pressureToJSON(const char *sensorName, pressure_t *input);
 
-/**
- * Write functions take care of persisting data to an SD card
- *
- * Data can be written in any format, including a custom binary format for
- * space efficiency, by using the writeBytes function that takes a byte array.
- *
- * The individual helpers write functions automate the process of saving data
- * for each individual value, but write in CSV format, which is easier to consume
- * after the experiment, but takes up more space.
- *
- * Finally, binaryWrite functions write optimized binary representations of the
- * data. This helps save SD card space, but means that the data must be decoded
- * after download before it can be used.
- */
-bool beginDataLog(int chipSelectPin, const char *fileNamePrefix, bool csvData);
-
-int writeBytes(const uint8_t *buffer, uint8_t numBytes);
-int writeString(const char *output);
-int writeAcceleration(const char *sensorName, acceleration_t *data);
-int writeMagnetic(const char *sensorName, magnetic_t *data);
-int writeGyro(const char *sensorName, gyro_t *data);
-int writeTemperature(const char *sensorName, temperature_t *data);
-int writeLuminosity(const char *sensorName, luminosity_t *data);
-int writeUVLight(const char *sensorName, uvlight_t *data);
-int writeOrientation(const char *sensorName, orientation_t *data);
-int writePressure(const char *sensorName, pressure_t *data);
-
-int binaryWriteAcceleration(const uint8_t sensorId, acceleration_t *data);
-int binaryWriteMagnetic(const uint8_t sensorId, magnetic_t *data);
-int binaryWriteGyro(const uint8_t sensorId, gyro_t *data);
-int binaryWriteTemperature(const uint8_t sensorId, temperature_t *data);
-int binaryWriteLuminosity(const uint8_t sensorId, luminosity_t *data);
-int binaryWriteUVLight(const uint8_t sensorId, uvlight_t *data);
-int binaryWriteOrientation(const uint8_t sensorId, orientation_t *data);
-int binaryWritePressure(const uint8_t sensorId, pressure_t *data);
-
-/**
- * Setup and use the RTC chip, if found
- *
- * The clock must be set before it can be used. This is achieved by compiling
- * and running a basic script (examples/set_rtc). After this, the library
- * automatically uses the RTC if available to timestamp.
- */
-bool setRTC();
-
 #ifdef __cplusplus
 } // extern "C"
 #endif
 
-#endif /* SENSORSDK_H_ */
+#endif /* ARDUSATSENSORS_H_ */
