+++ utility/drivers.cpp
@@ -12,7 +12,7 @@
 #include <Wire.h>
 #include "drivers.h"
 #include "pololu_LSM303.h"
-#include "ArdusatSDK.h"
+#include "ArdusatSensors.h"
 
 static config_lsm303_accel_t _lsm303_d_accel_config;
 static config_lsm303_mag_t _lsm303_d_mag_config;
